# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Contatos",
        "description": "Esta é a documentação da API contatos.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/people": {
            "get": {
                "tags": [
                    "Pessoas"
                ],
                "summary": "Lista todas as pessoas com seus contatos",
                "operationId": "e98ae4a18dc0a2ae0b94a6a59a39dc10",
                "responses": {
                    "200": {
                        "description": "Lista de pessoas com contatos"
                    },
                    "400": {
                        "description": "Erro ao processar a requisição",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao listar pessoas"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Pessoas"
                ],
                "summary": "Cria uma nova pessoa com contatos",
                "operationId": "5bd780ad283a1d97e0792b21f49a7416",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePersonRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pessoa criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao processar a requisição",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao criar pessoa"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/people/{person}": {
            "get": {
                "tags": [
                    "Pessoas"
                ],
                "summary": "Obtém detalhes de uma pessoa com contatos",
                "operationId": "975f1a51a61f7bf1855c1bde5b92a4c1",
                "parameters": [
                    {
                        "name": "person",
                        "in": "path",
                        "description": "ID da pessoa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da pessoa com contatos"
                    },
                    "400": {
                        "description": "Erro ao processar a requisição"
                    }
                }
            },
            "put": {
                "tags": [
                    "Pessoas"
                ],
                "summary": "Atualiza uma pessoa e seus contatos",
                "operationId": "41023ac18df385494b13a7044cff939f",
                "parameters": [
                    {
                        "name": "person",
                        "in": "path",
                        "description": "ID da pessoa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePersonRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pessoa atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao processar a requisição",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao atualizar pessoa"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pessoas"
                ],
                "summary": "Exclui uma pessoa e seus contatos",
                "operationId": "e6075bf0464a4e1d0cb907e31e50954c",
                "parameters": [
                    {
                        "name": "person",
                        "in": "path",
                        "description": "ID da pessoa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Pessoa excluída com sucesso"
                    },
                    "400": {
                        "description": "Erro ao processar a requisição",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao excluir pessoa"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "StorePersonRequest": {
                "required": [
                    "name",
                    "contacts"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Fulano de Tal"
                    },
                    "contacts": {
                        "type": "array",
                        "items": {
                            "required": [
                                "type",
                                "contact"
                            ],
                            "properties": {
                                "type": {
                                    "type": "string",
                                    "enum": [
                                        "phone",
                                        "email",
                                        "whatsapp"
                                    ]
                                },
                                "contact": {
                                    "type": "string",
                                    "example": "fulano@example.com"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "UpdatePersonRequest": {
                "required": [
                    "name",
                    "contacts"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Fulano de Tal"
                    },
                    "contacts": {
                        "type": "array",
                        "items": {
                            "required": [
                                "type",
                                "contact"
                            ],
                            "properties": {
                                "type": {
                                    "type": "string",
                                    "enum": [
                                        "phone",
                                        "email",
                                        "whatsapp"
                                    ]
                                },
                                "contact": {
                                    "type": "string",
                                    "example": "fulano@example.com"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
...
}